Gobal_Healthcare_Project/requirements.txt
mysql-connector-python
requests
pandas
tabulate